feat: CGYY version 1.2
Added support for proxy-ip when your account is banned by ip. While it can work, it's recommended that you use another network to snap up the court, not proxy, for its slow speed.
Make some small improvements on variable definition and logic. Added a timer option so that you can see how slow this script runs xd.
Still need improvement on speed, and logic for tests that are not in scheduled mode and after midnight, and support for freely choosing prior courts.
And most importantly, NEED A GUI.